Create Fake News Datasets on Facebook.xlsx
This thesis is based on “Fake News Detection on Facebook”. we concentrated entirely on Facebook. Because there had been no rich dataset on Facebook, we created a novel dataset (19 unique attributes) containing post reactions and page engagements data. Our primary goal was to develop a content-independent model that could detect serious fabrications and hoaxes of any kind on Facebook as it is more efficient than a content-dependent model.